LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/offset-distance-expected.txt


PASS Can set 'offset-distance' to CSS-wide keywords
FAIL Can set 'offset-distance' to var() references assert_equals: expected 2 but got 1
FAIL Can set 'offset-distance' to a length assert_class_string: specified calc must be a CSSMathSum expected "[object CSSMathSum]" but got "[object CSSUnitValue]"
FAIL Can set 'offset-distance' to a percent assert_class_string: specified calc must be a CSSMathSum expected "[object CSSMathSum]" but got "[object CSSUnitValue]"
PASS Setting 'offset-distance' to a time throws TypeError
PASS Setting 'offset-distance' to an angle throws TypeError
PASS Setting 'offset-distance' to a flexible length throws TypeError
FAIL Setting 'offset-distance' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'offset-distance' to a URL throws TypeError
PASS Setting 'offset-distance' to a transform throws TypeError
